+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+## [1.4.1] - 2018-02-16
+* [#33](https://github.com/Blackjacx/SHSearchBar/pull/33): Fix Pod Try To Quick Try Pod Example - [@blackjacx](https://github.com/blackjacx).
+* Add example for search bar in UINavigationBar - [@blackjacx](https://github.com/blackjacx).
+
 ## [1.4.0] - 2017-11-22
 ### Changed
 * [#31](https://github.com/Blackjacx/SHSearchBar/pull/31): Refactor Text Did Change Delegation - [@blackjacx](https://github.com/blackjacx).